backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx[socks]>=0.27.0
beautifulsoup4>=4.12.0
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-google-genai>=1.0.1
sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9
pytest>=8.3.0
apscheduler>=3.10.4
openai>=1.52.0
